Menambahkan validasi di tiap fitur

+++ app/Http/Controllers/dashboard/utama/randaKabilasa/MeningkatkanLifeSkillController.php
@@ -4,11 +4,15 @@ namespace App\Http\Controllers\dashboard\utama\randaKabilasa;
 
 use App\Http\Controllers\Controller;
 use App\Models\AnggotaKeluarga;
+use App\Models\Bidan;
 use App\Models\JawabanMeningkatkanLifeSkill;
+use App\Models\Pemberitahuan;
 use App\Models\RandaKabilasa;
 use App\Models\SoalMeningkatkanLifeSkill;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 
 class MeningkatkanLifeSkillController extends Controller
 {
@@ -29,9 +33,13 @@ class MeningkatkanLifeSkillController extends Controller
      */
     public function create(Request $request)
     {
-        $randaKabilasa = RandaKabilasa::find($request->randaKabilasa);
-        $daftarSoal = SoalMeningkatkanLifeSkill::orderBy('urutan', 'asc')->get();
-        return view('dashboard.pages.utama.randaKabilasa.meningkatkanLifeSkill.create', compact('randaKabilasa', 'daftarSoal'));
+        if (in_array(Auth::user()->role, ['admin', 'bidan', 'keluarga'])) {
+            $randaKabilasa = RandaKabilasa::find($request->randaKabilasa);
+            $daftarSoal = SoalMeningkatkanLifeSkill::orderBy('urutan', 'asc')->get();
+            return view('dashboard.pages.utama.randaKabilasa.meningkatkanLifeSkill.create', compact('randaKabilasa', 'daftarSoal'));
+        } else {
+            return abort(404);
+        }
     }
 
     /**
@@ -99,10 +107,19 @@ class MeningkatkanLifeSkillController extends Controller
     public function store(Request $request)
     {
         $data = $this->proses($request);
+
+        $role = Auth::user()->role;
         $randaKabilasa = RandaKabilasa::find($request->randaKabilasa);
 
         $randaKabilasa->is_meningkatkan_life_skill = 1;
         $randaKabilasa->kategori_meningkatkan_life_skill = $data['kategori'];
+
+        if ($role != 'keluarga') {
+            $randaKabilasa->is_valid_meningkatkan_life_skill = 1;
+        } else {
+            $randaKabilasa->is_valid_meningkatkan_life_skill = 0;
+        }
+
         $randaKabilasa->save();
 
         $soal = SoalMeningkatkanLifeSkill::orderBy('urutan', 'asc')->get();
@@ -142,6 +159,7 @@ class MeningkatkanLifeSkillController extends Controller
         $usia =  $interval->format('%y Tahun %m Bulan %d Hari');
 
         $data = [
+            'id' => $randaKabilasa->id,
             'nama_anak' => $anak->nama_lengkap,
             'tanggal_lahir' => Carbon::parse($anak->tanggal_lahir)->translatedFormat('d F Y'),
             'usia_tahun' => $usia,
@@ -150,7 +168,9 @@ class MeningkatkanLifeSkillController extends Controller
             'kategori' => $kategori,
             'tanggal_proses' => Carbon::parse($randaKabilasa->created_at)->translatedFormat('d F Y'),
             'tanggal_validasi' => Carbon::parse($randaKabilasa->tanggal_validasi)->translatedFormat('d F Y'),
-            'bidan' => $randaKabilasa->bidan->nama_lengkap,
+            'bidan' => $randaKabilasa->bidan->nama_lengkap ?? '-',
+            'is_valid_meningkatkan_life_skill' => $randaKabilasa->is_valid_meningkatkan_life_skill,
+            'bidan_konfirmasi' => $randaKabilasa->anggotaKeluarga->getBidan($randaKabilasa->anggota_keluarga_id)
         ];
 
         return view('dashboard.pages.utama.randaKabilasa.meningkatkanLifeSkill.show', compact('randaKabilasa', 'data', 'daftarSoal'));
@@ -165,8 +185,12 @@ class MeningkatkanLifeSkillController extends Controller
     public function edit(Request $request)
     {
         $randaKabilasa = RandaKabilasa::find($request->randaKabilasa);
-        $daftarSoal = SoalMeningkatkanLifeSkill::orderBy('urutan', 'asc')->get();
-        return view('dashboard.pages.utama.randaKabilasa.meningkatkanLifeSkill.edit', compact('randaKabilasa', 'daftarSoal'));
+        if ((Auth::user()->profil->id == $randaKabilasa->bidan_id) || (Auth::user()->role == 'admin') || (Auth::user()->profil->kartu_keluarga_id == $randaKabilasa->anggotaKeluarga->kartu_keluarga_id)) {
+            $daftarSoal = SoalMeningkatkanLifeSkill::orderBy('urutan', 'asc')->get();
+            return view('dashboard.pages.utama.randaKabilasa.meningkatkanLifeSkill.edit', compact('randaKabilasa', 'daftarSoal'));
+        } else {
+            return abort(404);
+        }
     }
 
     /**
@@ -184,6 +208,15 @@ class MeningkatkanLifeSkillController extends Controller
         $randaKabilasa = RandaKabilasa::find($request->randaKabilasa);
 
         $randaKabilasa->kategori_meningkatkan_life_skill = $data['kategori'];
+
+
+        if ((Auth::user()->role == 'keluarga') && ($randaKabilasa->is_valid_meningkatkan_life_skill == 2)) {
+            $randaKabilasa->is_valid_meningkatkan_life_skill = 0;
+            $randaKabilasa->bidan_id = null;
+            $randaKabilasa->tanggal_validasi = null;
+            $randaKabilasa->alasan_ditolak_meningkatkan_life_skill = null;
+        }
+
         $randaKabilasa->save();
 
         $jawabanMeningkatkanLifeSkill = JawabanMeningkatkanLifeSkill::where('randa_kabilasa_id', $request->randaKabilasa)->delete();
@@ -195,6 +228,14 @@ class MeningkatkanLifeSkillController extends Controller
             $jawabanMeningkatkanLifeSkill->soal_id = $soal[$i]->id;
             $jawabanMeningkatkanLifeSkill->jawaban = $request->$jawaban[0];
             $jawabanMeningkatkanLifeSkill->save();
+        }
+
+        $pemberitahuan = Pemberitahuan::where('anggota_keluarga_id', $randaKabilasa->anggota_keluarga_id)
+            ->where('tentang', 'meningkatkan_life_skill')
+            ->where('fitur_id', $randaKabilasa->id);
+
+        if ($pemberitahuan) {
+            $pemberitahuan->delete();
         }
 
         return response()->json([
@@ -211,14 +252,89 @@ class MeningkatkanLifeSkillController extends Controller
     public function destroy(Request $request)
     {
         $randaKabilasa = RandaKabilasa::find($request->randaKabilasa);
-        $randaKabilasa->is_meningkatkan_life_skill = 0;
-        $randaKabilasa->kategori_meningkatkan_life_skill = null;
+        if ((Auth::user()->profil->id == $randaKabilasa->bidan_id) || (Auth::user()->role == 'admin')) {
+            $randaKabilasa->is_meningkatkan_life_skill = 0;
+            $randaKabilasa->kategori_meningkatkan_life_skill = null;
+            $randaKabilasa->save();
+
+            $jawabanMeningkatkanLifeSkill = JawabanMeningkatkanLifeSkill::where('randa_kabilasa_id', $request->randaKabilasa)->delete();
+
+            $pemberitahuan = Pemberitahuan::where('fitur_id', $randaKabilasa->id)
+                ->where('tentang', 'meningkatkan_life_skill');
+
+            if ($pemberitahuan) {
+                $pemberitahuan->delete();
+            }
+
+            return response()->json([
+                'status' => 'success'
+            ]);
+        } else {
+            return abort(404);
+        }
+    }
+
+    public function validasi(Request $request, RandaKabilasa $randaKabilasa)
+    {
+        $id = $request->id;
+
+        if ($request->konfirmasi == 1) {
+            $alasan_req = '';
+            $alasan = null;
+        } else {
+            $alasan_req = 'required';
+            $alasan = $request->alasan;
+        }
+
+        if (Auth::user()->role == 'admin') {
+            $bidan_id_req = 'required';
+            $bidan_id = $request->bidan_id;
+        } else {
+            $bidan_id_req = '';
+            $bidan_id = Auth::user()->profil->id;
+        }
+
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'bidan_id' => $bidan_id_req,
+                'konfirmasi' => 'required',
+                'alasan' => $alasan_req,
+            ],
+            [
+                'bidan_id.required' => 'Bidan harus diisi',
+                'konfirmasi.required' => 'Konfirmasi harus diisi',
+                'alasan.required' => 'Alasan harus diisi',
+            ]
+        );
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()]);
+        }
+
+        $randaKabilasa->is_valid_meningkatkan_life_skill = $request->konfirmasi;
+        $randaKabilasa->bidan_id = $bidan_id;
+        $randaKabilasa->tanggal_validasi = Carbon::now();
+        $randaKabilasa->alasan_ditolak_meningkatkan_life_skill = $alasan;
         $randaKabilasa->save();
 
-        $jawabanMeningkatkanLifeSkill = JawabanMeningkatkanLifeSkill::where('randa_kabilasa_id', $request->randaKabilasa)->delete();
+        $namaBidan = Bidan::where('id', $bidan_id)->first();
 
-        return response()->json([
-            'status' => 'success'
-        ]);
+        $pemberitahuan = new Pemberitahuan();
+        $pemberitahuan->user_id = $randaKabilasa->anggotaKeluarga->kartuKeluarga->kepalaKeluarga->user_id;
+        $pemberitahuan->fitur_id = $randaKabilasa->id;
+        $pemberitahuan->anggota_keluarga_id = $randaKabilasa->anggota_keluarga_id;
+        $pemberitahuan->tentang = 'meningkatkan_life_skill';
+
+        if ($request->konfirmasi == 1) {
+            $pemberitahuan->judul = 'Selamat, data asesmen meningkatkan life skill dan potensi diri anda telah divalidasi.';
+            $pemberitahuan->isi = 'Data asesmen meningkatkan life skill dan potensi diri anda (' . ucwords(strtolower($randaKabilasa->anggotaKeluarga->nama_lengkap)) . ') divalidasi oleh bidan ' . $namaBidan->nama_lengkap . '.';
+        } else {
+            $pemberitahuan->judul = 'Maaf, data asesmen meningkatkan life skill dan potensi diri anda' . ' (' . ucwords(strtolower($randaKabilasa->anggotaKeluarga->nama_lengkap)) . ') ditolak.';
+            $pemberitahuan->isi = 'Silahkan perbarui data untuk melihat alasan data asesmen meningkatkan life skill dan potensi diri ditolak dan mengirim ulang data. Terima Kasih.';
+        }
+
+        $pemberitahuan->save();
+        return response()->json(['res' => 'success', 'konfirmasi' => $request->konfirmasi]);
     }
 }
